# contract/openapi.v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "testhost | v1",
    "description": "",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost/"
    }
  ],
  "paths": {
...
    "/api/v1/t/{tenantSlug}/roster/{rosterEntryId}/rank": {
      "put": {
        "tags": [
          "Roster"
        ],
        "parameters": [
          {
            "name": "rosterEntryId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "tenantSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetRosterRankViewModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SetRosterRankViewModel"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/SetRosterRankViewModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/t/{tenantSlug}/roster/{rosterEntryId}/note": {
      "put": {
        "tags": [
          "Roster"
        ],
        "parameters": [
          {
            "name": "rosterEntryId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "tenantSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetOfficerNoteViewModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SetOfficerNoteViewModel"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/SetOfficerNoteViewModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/t/{tenantSlug}/roster/{rosterEntryId}": {
      "delete": {
        "tags": [
          "Roster"
        ],
        "parameters": [
          {
            "name": "rosterEntryId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "tenantSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/t/{tenantSlug}/roster/{rosterEntryId}/main": {
      "put": {
        "tags": [
          "Roster"
        ],
        "parameters": [
          {
            "name": "rosterEntryId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "tenantSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LinkAltViewModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LinkAltViewModel"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/LinkAltViewModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Roster"
        ],
        "parameters": [
          {
            "name": "rosterEntryId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "tenantSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/t/{tenantSlug}/guilds/{guildId}/sync": {
      "post": {
        "tags": [
          "TenantGuilds"
        ],
        "parameters": [
          {
            "name": "guildId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "tenantSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GuildServiceModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuildServiceModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuildServiceModel"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/t/{tenantSlug}/guilds/{guildId}": {
      "delete": {
        "tags": [
          "TenantGuilds"
        ],
        "parameters": [
          {
            "name": "guildId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "tenantSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CharacterFaction": {
        "enum": [
          "Alliance",
          "Horde"
        ]
      },
...
      "GuildServiceModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "realmSlug": {
            "type": "string"
          },
          "faction": {
            "$ref": "#/components/schemas/CharacterFaction"
          },
          "memberCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "lastSyncedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "LinkAltViewModel": {
        "type": "object",
        "properties": {
          "mainRosterEntryId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "SetOfficerNoteViewModel": {
        "type": "object",
        "properties": {
          "officerNote": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "SetRosterRankViewModel": {
        "type": "object",
        "properties": {
          "tenantRankId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          }
        }
      },
...
    }
  },
...
}